requirements.txt
fastapi
uvicorn[standard]
pydantic
httpx
openai
python-dotenv
langchain
langchain-openai
langgraph
joblib
pandas
geopandas
numpy
scikit-learn
xgboost
shap
rasterio
xarray
requests
